requirements.txt
# 투자 데이터 분석 프로젝트 의존성
# pip install -r requirements.txt

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# 주가 수집
yfinance>=0.2.40

# 경제지표 수집
fredapi>=0.5.0
requests>=2.31.0

# 시각화
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
mplfinance>=0.12.10

# 대시보드
streamlit>=1.28.0

# InfluxDB 연동
influxdb-client>=1.38.0

# 환경변수 관리
python-dotenv>=1.0.0

# 감성분석 (Phase 3, 선택)
# konlpy>=0.6.0
# transformers>=4.35.0
